+++ src/sdk/moduleconfig.h
@@ -15,7 +15,7 @@
  *   to add multiple entries.
  */
 #define MODULE_NAME "KTP CURL AMXX"
-#define MODULE_VERSION "1.3.12-ktp"
+#define MODULE_VERSION "1.3.13-ktp"
 #define MODULE_AUTHOR "Polarhigh / KTP"
 #define MODULE_URL "https://github.com/afraznein/KTPAMXXCurl"
 #define MODULE_LOGTAG "CURL"